add directory core

+++ db/models.py
@@ -4,14 +4,14 @@ class CarItem(BaseModel):
     id: int
     name: str
     description: str
-    price: float
+    price: int
     available: bool
 
 class Order(BaseModel):
     id: int
     car_id: int
     user_id: int
-    total_price: float
+    total_price: int
